id_sd_opl2alsa: Add support for real OPL2s on Linux
This adds support for using real Adlib/Sound Blaster/OPL2-based
soundcards under Linux, by using ALSA's FM synth support. OPL2 register
writes are converted to ALSA ioctls(), which are then forwarded to the
card by the kernel.

This should work out-of-the-box on most Linux systems with a real OPL2
(though these are rare). Unlike ioperm()-based solutions, this does not
require root access, and works automatically with soundcards which use
non-standard IO port addresses (e.g. all PCIe cards).

This can be built by passing WITH_ALSA=1 to 'make'. The ALSA development
packages for your distro must be installed.
It can then be activated with the /ALSAOPL command-line option.

The hardware device defaults to "hw:0,0", but can be overridden with the
/ALSADEV <device> argument. This should be the 'hwdep' ALSA device for
the FM synth.

Note that, because ALSA doesn't expose the PC Speaker, it is emulated
(quite badly) with the OPL2 if enabled.

Note that I've only tested this on a CMI8378 PCIe card, but it worked
perfectly out-of-the-box with that. :-)