id_sd_opl2lpt: Add support for the OPL2LPT via libieee1284
This change adds support for the OPL2LPT parallel-port soundcard:
https://www.serdashop.com/OPL2LPT
It should also work with the OPL3LPT, but I haven't tried this yet!

It uses libieee1284 to provide access to the parallel port, and
Adlib/OPL2 register writes are redirected to that port.

You can build this by enabling the WITH_IEEE1284=1 option when calling
'make'. It must then be activated at runtime with the /OPL2LPT option.
Note that this uses libieee1284, so you'll need the development packages
for your distro to use this.

This code will always use the first available parallel port (typically
"parport0"). Note that if there's another driver using the parallel
port, Omnispeak will not be able to access it. This is usually the case
on Linux systems, where the 'lp' kernel driver is often active for all
parallel ports. Running 'sudo rmmod lp' is usually sufficient. You'll
also need to make sure that your user has the correct permissions for
the parallel port: if they do, Omnispeak need not be run as root.

(Note that there is some code present for using ioperm() to write to the
parallel port directly, bypassing the kernel. This is incomplete, but
can be made to work. There are several problems with it, though, so I'd
recommend using libieee1284 instead.)

The PC-speaker is emulated the same way as with the opl2alsa backend. It
sounds horrible, so you really should use Adlib sound effects if you've
got an OPL2LPT.