id_fs: Make default paths configurable, support XDG
This is a significant change to make much of the filesystem config
defaults configurable at compile time. Hopefully this will help
packagers build omnispeak without the need for patches.

In addition, it adds support for using the XDG basedir specification[1]
to determine the "User" path (for savegames, etc). By default, this is
used as a fallback if the default "User" path (normally the current
directory) is not writable, but this can be overridden with the
XDGUSERDIR=1 makefile option (or the FS_USER_PATH_PREFER_XDG #define in
ck_config.h).

In addition, a warning is printed if the final User path is not
writable.

More details can be found in ck_config.h, which has the defaults all
specified. 'make help' will list the Makefile options that can be used
(which is a subset of what's exposed in ck_config.h).

[1]: https://specifications.freedesktop.org/basedir-spec/basedir-spec-latest.html