id_sd_sdl: Add support for the Nuked OPL3 emulator
This was requested by aroenai in issue #30.

The Nuked OPL3 emulator is a more accurate emulator of OPL3 chips. It
can now be used in place of the DBOPL emulator from DOSBox by passing
the /NUKEDOPL3 option to Omnispeak.

There are still a few things I'd like to fix with this (and that switch
may change), so documenting it properly will wait for another commit.

The biggest issue is that Nuked OPL3 emulates an OPL3 (which is stereo),
whereas Keen was written against an OPL2 (which is mono). Omnispeak's
sound code therefore only supports monaural sound, and we have to mix
the OPL3's stereo output down before outputting it. This mixing is very
dodgy (I'm not even clamping the results to 16-bit properly, so clicks,
pops, and crackle on loud sounds may be possible.

Note also that Nuked OPL3 is under the LGPL 2.1, whereas the rest of
Omnispeak (including the DOSBox bits) are under the GPL v2. This is
almost certainly not going to be a problem, but there's a separate
NUKEDOPL3-LICENSE file included in the src/opl/ directory nevertheless.