Makefile: Default to the SDL2 renderer on Windows
Windows is much less likely to have a working OpenGL implementatin
(particularly with OpenGL 2.0 support), and SDL has several backends
which work better under Windows. This makes up for the fact that the
SDL2 renderer backend does palette conversions in software.

And with this, the Windows builds run out-of-the-box on a WinXP VM with
no 3D graphics drivers, so that's nice.

OpenGL builds for windows can still be made by setting RENDERER=sdl2gl.