ci: Build Linux versions with ALSA & IEEE1284
Unfortunately, the libieee1284 dependency may cause the binaries not to
work on everyone's machine out of the box, but it's worth at least
building them in CI to make sure the code does build.

Ideally, maybe we'd have some proper packaging code and upload nicer
packages as the artefacts, but for now, this is it.