id_sd_sdl: Small fixes for Nuked OPL3 support
This change clamps the mixed output from the Stereo→Mono downmix, which
ought to limit the nastiness of any crackling, and bring this up to par
with the DBOPL backend.

It also adds support for very high sample rates (which don't ever occur
in practice) by allowing a mix length of greater than 512. If this is
the case, the sound is split into chunks for mixing. (It doesn't really
matter, but it makes me sleep better at night.)

If we're luckly, no-one will ever notice these changes, as the issues
they fix may never actually occur. But I can't prove they won't, so here
we are.