id_cfg: Add a new config file system
The CONFIG.CKx files which Keen normally uses to store settings have a
few flaws:
- They're not really extensible to new settings.
- They're not human readable / modifiable.
- They're tied to the particular game (so settings aren't shared between
  differen episodes).

Because of this, Omnispeak-specific settings (like fullscreen/windowed
mode, etc) are not saved anywhere at all.

This change adds a brand-new configuration file system, ID_CFG. It
supports both integer and string settings, which are stored in a simple
key/value file called OMNISPK.CFG.

The CFG_GetConfigInt() and CFG_GetConfigString() functions can be used
to read integer and string config options respectively. They also expect
a "default" value, which is returned if the config item is not found in
the file.

The CFG_SetConfigInt() and CFG_SetConfigString() function allow the
value of these config options to be set. When the game shuts down,
CFG_SaveConfig() will update the OMNISPK.CFG file with the new values,
or add a new line to the file as necessary. Formatting and comments
already in the file should be preserved.