ck6_obj1: Fix the portraits when Keen doesn't have map items
The dialog boxes for when Keen tried to climb the cliff, get past the
grabbitter, and use the rocket had misaligned Keen portraits, due to VH_
being used instead of VHB_. We fixed this for the dialogs in ck6_misc.c
when dirty-rectangle drawing was enabled, but missed these ones. (I
guess I never tried climbing the cliff without the rope. Whoops!)