id_vl_dos: Make double-buffering / scrolling more efficient/accurate
As of this patch, double-buffering in the VL_DOS backend should match
the way Keen works (in SVGA compatibility mode) much more closely.

In particular, instead of both pages being part of the one large
allocation, and being moved around (according to some dodgy heuristics)
as one big block when we're about to scroll out of the video memory
bounds, each page is now moved independently. This amortizes the slow
SVGA compatibility copy over a couple of frames: instead of copying both
at once and having one very slow frame (with visible glitches), we now
have two slow frames (but half as slow), spread out over a longer
scroll, which half the time will be of the offscreen backbuffer page, so
not causing glitches.

This requires keeping a gap between the buffers, which it turns out Keen
actually does (neat!)

Note that this isn't finished or perfect yet: we don't update the CRTC
scanout offset immediately after copying. Plus VSync is still very
dodgy, which also causes some artefacts.

But it's an improvement, so that's nice.