id_vl_dos: Don't allocate twice as much memory as we need
For system memory surfaces (everything except the framebuffer), we were
allocating w*h bytes, but we actually only need half that, as the images
are 4 bpp, not 8 bpp (unlike the other backends).

This stems from an early experimental mode 13h dos/vga backend, which
was inefficient, and so slow as to be unusable.