id_fs: Define _GNU_SOURCE in order to get O_PATH for Linux backend
The "linux" id_fs backend also needs _GNU_SOURCE to be #defined before
any system headers are included. Otherwise, O_PATH isn't defined, which
we use when creating new files. This is actually documented in the man
page for 'openat', so I really should've done this before, but it worked
in C++ for some ill-defined reason.

Probably the cleverer thing would be to drop the dependency on O_PATH,
but for now, this is it.