id_cfg: Handle failure to open config file for writing
In CFG_SaveConfig(), we'd try to write to and close a config file, even
if we were unable to open it succesfully. This change handles the case
where the file cannot be saved more gracefully (printing an error).

This could occur if, for example, the user path was set incorrectly.