ck_play: Refactor Sprite Test to not use goto
The CK_SpriteTest() function used a somewhat ugly goto statement, which
was also annoying the Watcom compiler. This change reworks it to have
the redraw code part of the inner loop, but only trigger if the chunk or
shift has changed. It seems to work pretty well here, and with this
change ck_play.c compiles under Open Watcom v2.