id_mm: Support malloc(0) returning NULL
On some CRT implementations, malloc(0) will return a NULL pointer. This
is perfectly valid, but our ID_MM implementation would assume failure if
this occurred. This would cause MM_GetPtr() to repeatly error about
out-of-memory.

This change only treats malloc(0) returning NULL as an error if size is
nonzero. ID_MM will forward the NULL pointer on to the rest of the
program, but since it's illegal to ever access any bytes of a 0-sized
allocation (well, there aren't any bytes to access, so…), this should
never cause any problems.

Maybe there could be some code which checks NULL to see if a block has
been purged, which could be a problem, though. I'll need to look into
that. In the meantime, this fixes issues with Watcom's slightly dodgy
runtime implementation.