id_sd_opl2alsa: Use structs correctly under C
A few of the ALSA structs — snd_dm_fm_{voice,note,params} — were being
accessed directly as types, rather than with the 'struct' keyword. This
works under C++, and as we set BUILDASCPP=1 by default on Linux I hadn't
noticed.

With this fixed, we can build with WITH_ALSA=1 and BUILDASCPP=0.