CMakeLists: Add a CMake-based build system
Note that the Makefile in src/ is still the recommended way to build
Omnispeak. However, it is very gcc-specific. This CMakeLists.txt file
provides a somewhat broken subset of its functionality, but can (with
some effort) be made to build on non-gcc-compatible compilers.

To use it, create a build directory, and run something like:
	cd build/
	cmake .. -DRENDERER=sdl2 -DWITH_ALSA=1
	cmake --build .

Other build systems can be used by passing -G <buildsystem> to the first
CMake command.

Note that it doesn't copy data files around, so you'll need to do that
manually, or else use /GAMEPATH to point Omnispeak at them.