CI: Add a windows-based builder using MS Visual C++
This change adds a new GitHub CI workflow for building Win64 builds with
MSVC and the new CMake build system. It's very basic, and quite broken
in a number of ways, but should be a good start for making this codebase
more portable.

Note that CMake and SDL don't play quite as nicely together yet as I'd
like: in order for find_package(SDL2) to work, we manually set SDL2_DIR
_and_ copy an sdl2-config.cmake file I randomly found on the internet
into a copy of the MSVC dev binaries from the SDL website. Yuck! But it
works. There's some discussion about how to handle this better on the
SDL mailing lists:
https://discourse.libsdl.org/t/how-is-sdl2-supposed-to-be-used-with-cmake/31275

Note that the generated binaries aren't too useful yet, as they need a
lot of MSVC runtime files to run. They're also all 64-bit: I'm sure
there's a way to force a 32-bit build, but I haven't worked it out yet.

Finally, MSVC generates a lot of compile warnings. Most of these are
about insecure functions, and are advertising the use of Microsoft's
'secure CRT' replacements. I'm leaving these for now, as it's probably
worth doing an audit of the CRT functions we're using for obvious buffer
overflows, etc, anyway.