Makefile: Use --std=gnu99 for BUILDASCPP=0
It turns out that djgpp doesn't provide any of <dirent.h> if
__STRICT_ANSI__ is defined, which is the case when --std=c99 is set.
This meant that BUILDASCPP=0 has been broken for DOS basically forever.

It also makes sense, as the directory functions aren't (or at least
weren't) technically a part of ANSI C, but POSIX extensions. Of course,
basically everyone supports them…

It would be possible to get rid of the whole opendir() bit on DOS, I
guess, as case sensitivity isn't a problem (and it's the only thing we
really _need_ opendir() for). Maybe we'll try that later…