id_rf: Clean up RF_ReplaceTileBlock()
This function previously used a lot of manual pointer arithmetic, which
had been half-replaced with repeated manual references to the tile data.
A number of the variables being set, used, and updated were then not
used.

I started replacing this with CA_TilePtrAtPos(), but realised pretty
quickly that we could get rid of all the pointers, and just use
CA_TileAtPos() and CA_SetTileAtPos(). This is much nicer, IMHO, and has
a couple fewer warnings with -Wall.