ck_inter: Fix clearing of terminator pics in ScrollTerminatorCredits
I think this now correctly clears just the necessary bits off the
scrolling credits, so we no longer need the hacky "just clear
everything" code. Hopefully, this means the code will be faster, as well
as more accurate.

The complication here is that the original clear code cleared a 16-bit
word at a time (using rep stosw), and caluclated a "clear width" based
on that.

This also helps clean up some unused-but-set variable warnings.