id_cfg: Add a new Boolean config type.
Instead of directly using integers for boolean config options, we now
use the bool type. Integer and bool values are largely interchangeable,
but bools will always be saved as "true" / "false", and "true" / "false"
will be loaded correctly.

These are implemented as integers under the hood with an extra
'is_boolean' flag.