ck_play: Fix playloop dumper compilation
The playloop dumper did not compile anymore, as it still used
CK_Cross_fwriteInt32LE(), which was replaced with ID_FS functions.

Note that technically, this is still invalid, as we're using
FS_WriteInt32LE() on a FILE* which isn't opened by ID_FS. This is to
make the dumpfile appear in the current directory, where it's expected
to be. I may or may not change this at some point.