ck_main: Add /PLAYDEMO option to play a built-in demo
We already have the /DEMOFILE option to play a demo which is saved to an
external file, but no way of playing an individual internal demos.

The new /PLAYDEMO <n> option will play demo number <n>, then quit. The
goal of this is to be able to combine it with /DUMPFILE to get a dump
associated with a demo.