ck4_obj3: Cancel Mimrock x velocity on hitting right TI, not top TI
Whem Mimrocks bounce, their x velocity should be cancelled if they've
hit a wall, but instead only was cancelled for leftTI walls, or floors
(topTI). This meant that it was not being cancelled in some cases, and
incorrectly cancelled when hitting a floor.

The latter was causing the dump for Demo 4 (the high scores) to not
match. It doesn't appear to have had any lingering effect, however.