ck_misc: Hack "correct" dump values for ShrapnelTileCol
The Shelley in the Keen 5 High Scores explodes, and a bad action is set
on the Shelley Bits in ShrapnelTileCol. This is not properly removed
immediately, and the dump mismatched as a result. We had some variables
set for dump accuracy, here we add some more for the Demo dump.

Note that the '3' for actionTimer in particular really annoys me, but
we'll keep it for now as it fixes _all_ the demo dumps across 4, 5, and
6v14.