ck_keen: Add a config option to disable Impossible Bullet
The "Impossible Bullet" bug in Keen 6 is just that: a bug. Add an option
to fix it. In OMNISPK.CFG, just add
	impossibleBullet = false

It's set to true (to enable impossible bullet) by default, which matches
the original game.

Note that enabling this does not add the bug to Keen 4 and 5, nor does
this stop you from using it if you ride a platform off the bottom of the
map. It stops Keen from landing on his own bullets, not from riding them
afterwards.