ck4_obj1: Fix bird type on egg hatching
We'd forgotten to set the Bird object's type correctly when it hatched
from an egg, which could lead to a "FullClipToWalls: Bad obclass" error
after the bird takes off.

This confusion probably occured because of the CK4_HatchEgg() function
was so named, when it's actually used when the bird is recovering from
being stunned, so rename it. (The bird is actually a separate object
from the egg.)