ck4_obj1: Birds should only follow keen when "free"
Thanks to K1n9_Duk3 for finding this!

In CK4_BirdFly (the think function for flying birds), the Bird should
only attempt to follow keen if user1 is 0, not non-zero. The user1
variable is set if the bird is colliding on the top or bottom.

This basically stops a bird from hovering under a small platform if Keen
is standing above it (and, in theory, vice-versa, though the Bird will
usually land). Insead, the bird will not repeatedly turn around, but
will instead keep going in its current direction until it can once again
move vertically, allowing it to navigate around small platform
obstacles, such as the ones in K4L5 (Chasm of Chills).