id_in: Store custom joystick options in the OMNISPK.CFG file
These are no-longer episode-specific, but the modern mode and deadzone
are actually saved.

Note that this is not how we'll probably want to implement this:
there's not much point having the IN_JoyConf* stuff at all rather than
just directly reading and setting the config variables.

Also, this totally breaks the defaults: everything's set to 0, so that
needs fixing.