id_us_1: Support overriding control type from config
Make it possible to force in_controlType to be stored in OMNISPK.CFG
instead of CONFIG.CKx.

If an in_controlType entry exists (and is not set to -1), it will be
used and updated as the control type. Currently they're just integer
values for the raw control type. 0 is keyboard, 2 is joystick 1, 3 is
joystick 2.

This means that the config will persist across episodes. Note that, at
present, the config will be saved to both CONFIG.CKx and OMNISPK.CFG if
enabled.