ck_game: Submit high scores on QED destruction (Thanks @LoneFox78)
The call to CK_SubmitHighScore() when the level state is LS_DestroyedQED
(ex_qedbroke) was commented out for some reason. This meant that, after
the end-game text in Keen 5, the high scores were shown without giving
the player a chance to submit theirs.

Removing the comment does indeed fix bug #38. Thanks to @LoneFox78 for
finding this and tracking down the root cause.