id_in: Fix an integer overflow bug with "modern" joy controls
When the joystick is pressed to the top-left corner, SDL reports the
axes as (-32768,-32768), which causes an integer overflow in the modern
joystick controls' deadzone detection code.

This is because:
- Signed integers can represent one more negative value than "positive"
  values (because 'zero' is 'positive').
- So we take the absolute value, which no longer fits in an int16.
- When we calculate the squared vector length, we hit an integer
  overflow and end up with a negative number.
- Because a negative distance from the centre is within the deadzone, as
  it were, the input is treated as zero.

This affected both SDL 1.2 and SDL 2.0. Haven't tested it with others.

We fix it by just clamping the input to 32767 at the appropriate point.
Alternatively, we could redo it to cast to unsigned, which might be
nicer, but this works for now.