id_in: Support joysticks in the main menu, even when not selected
It's a huge pain to not be able to enable a joystick without first using
the keyboard to nagivate the menu.

Add a new config option, in_forceJoyMenu, which makes the joystick
always work in the main menu, even if it's not the selected control
type. (This defaults to the first joystick for now, but maybe could be
made to loop over all of them later.)

This option is enabled by default, except in vanilla builds.