id_ca: Make CAL_Read[U][D]Word() less alignment-sensitive
The CAL_ReadWord() family of functions simply read a word from the
pointer using a cast, and caused it to crash on architectures which
don't support reading unaligned integers. Instead, this memcpy()s the
word in question into an aligned (stack) temporary, which is then
endian-converted.

Hopefully this renders @Ryzee119's hack here unnecessary:
https://github.com/sulix/omnispeak/commit/1683e665e4b4b2d8fc675460b2c06a184b3af9be