id_rf: Correct RF_MAX_SPRITETABLEENTRIES to 60 (Fix #42)
We had limited the concurrent onscreen sprites to 48, which
(surprisingly) worked in almost all cases, but Keen actually limits this
to 60. This caused an error in Gravitational Damping Hub (Keen5Level11)
with Vitalin shadows (bug #42).

I verified this against K1n9_Duk3's reconstructed code: MAXSPRITES in
ID_RF.H is 60 for all three episodes. There are a couple of other limits
which we have which don't match (particularly for things like max anim
timers, which vary between episodes), but our values are greater, so it
doesn't cause problems.