id_vl: Initial support for VSync
This change adds a new config option, "swapInterval", which is exposed
as "vsync" in the config menu. It controls the OpenGL swap interval in
the sdl2gl backend, and SDL_RENDERER_PRESENTVSYNC in the sdl2 backend.

It can be toggled at runtime with OpenGL, but not with SDL2, where its
value at startup is maintained. (The menu looks like it works, but
doesn't).

The /VSYNC and /NOVSYNC command-line options also control this variable.

Adaptive vsync will work (but cannot be toggled from the UI): set
swapInterval=-1 in OMNISPK.CFG