id_vl: Initial support for /NOPAN parameter
This rather useless option disables smooth scrolling entirely, and does
some small amount of work to adjust other parts of the gameplay code
(notably the keen-riding-platform bits) to make it at least look stable.

I think there might still be a bug or two in this, and the comments are
probably missing something, but it's worth having it out there a bit.
Equally the DOS/EGA backend should not write the horizontal PEL PANNING
register _at all_, rather than writing zero to it, and maybe we should
add this as a OMNISPK.CFG option as well.

But for now… enjoy(?), I guess…