ck_keen: Be pushed by Giks (and Flect and Blooglet) in SpecialColFunc
The fact that Flects and Blooglets do this (specifically in the Y
direction) might be the cause of Blooglet riding, but I haven't actually
checked:
https://pckf.com/viewtopic.php?t=11957

We'd missed the call to CK_PhysPushY() as well, which is interesting, as
Keen 5 doesn't do any such thing for Amptons and Koraths, and Keen 4
uses CK_PhysPushXY() for Bounders. This delightful inconsistency is no
doubt to keep us on our toes.