id_cfg: Don't shutdown if we never started up.
Sometimes Quit() would be called before all of the ID managers
would be started up. In the case of ID_CFG, this resulted in an
attempt to save a nonexistent config (which would crash trying
to look things up in the nonexistent hashtable), and potentially
a memory error trying to destroy a nonexistent arena.

This could be reproduced by passing an NULL userpath in:
	./omnispeak /USERPATH
(And hence was found as a result of looking into issue #43)