id_us: US_CheckParm: Treat "" as list terminator (Fix #43)
Thanks to @LoneFox78 for finding this issue.

The vl_parmStrings[] array was terminated with an empty string,
not a NULL pointer. This could cause US_CheckParm() to read past
the end of the array, as it ignored empty strings, but did not
treat them as a terminator.

The original game, however, did use "" as a terminator for its
equivalent of vl_parmStrings, and US_CheckParm() treated it the
same way as it treated NULL. We should do the same, rather than
changing the terminator in ID_VL.

AddressSanitizer shows the following error:
==18985==ERROR: AddressSanitizer: global-buffer-overflow on address 0x55b291d96198 at pc 0x55b291c9441a bp 0x7fffa349a080 sp 0x7fffa349a070
READ of size 8 at 0x55b291d96198 thread T0
    #0 0x55b291c94419 in US_CheckParm(char const*, char const**) (/home/david/Development/omnispeak/bin/omnispeak+0x50419)
    #1 0x55b291ca7308 in VL_Startup() (/home/david/Development/omnispeak/bin/omnispeak+0x63308)
    #2 0x55b291d3f40f in CK_InitGame() (/home/david/Development/omnispeak/bin/omnispeak+0xfb40f)
    #3 0x55b291d40175 in main (/home/david/Development/omnispeak/bin/omnispeak+0xfc175)
    #4 0x7fb36fb30d8f in __libc_start_call_main ../sysdeps/nptl/libc_start_call_main.h:58
    #5 0x7fb36fb30e3f in __libc_start_main_impl ../csu/libc-start.c:392
    #6 0x55b291c7d074 in _start (/home/david/Development/omnispeak/bin/omnispeak+0x39074)

0x55b291d96198 is located 0 bytes to the right of global variable 'vl_parmStrings' defined in 'id_vl.c:721:20' (0x55b291d96180) of size 24
SUMMARY: AddressSanitizer: global-buffer-overflow (/home/david/Development/omnispeak/bin/omnispeak+0x50419) in US_CheckParm(char const*, char const**)

Fixes: c505d6b4db ("id_vl: Initial support for /NOPAN parameter")