id_rf: Allow configuring rf_minTics
RF_CalcTics(), which works out the in-game timing, limits each frame to
run in a minimum of MINTICS tics (which, at the 70 Hz VGA ticrate, equals a
framerate of 35 Hz).

Previously this was hardcoded to 2, like in the original game. Setting
it to 1 would allow the game to run at 70 fps, but the physics feel
noticably different and some things (ledge grabbing) are a bit harder.

(DEMO recording and playback is always done at 3 tics per frame. Normal,
non-DEMO play runs as fast as it can, between 2 and 5 tics/frame.)

Allow MINTICS to be set as an OMNISPK.CFG option, to allow people to try
the "smoother" physics if they prefer.