id_ca: Give a (slightly) better message when expanding bad chunks (#44, #45)
If the EGAHEAD and EGAGRAPH files don't match (which will often happen
if the wrong version of Keen is used with the provided graphics files),
the chunk lengths will typically be totally wrong. Attempting to
decompress them leaves us with a terrible error message, so provide a
(slightly) better one, letting us know that it's chunk expansion which
is failing, and suggesting that it's a version mismatch.

This is not the ideal long-term solution for this, but the extra
sanity-checking doesn't hurt, so we'll do it anyway.

Thanks @synthead for reporting this.