id_ca: Harden CA_HuffExpand() and uses against bad data
CA_HuffExpand() and several functions which feed into it now accept the
compressed length as well as the expanded length, and will refuse to
read past the end of the input. This prevents a number of segfaults on
totally bad data.

Most notably, Keen 6 v1.4 data with v1.5 headers should now survive long
enough to print a warning message.