id_vl_sdl2vk: Rebuild SPIR-V shaders during build process
Instead of having hardcoded, binary SPIR-V shaders which were built a
very long time ago (and hence actually cause validation errors), include
the SPIR-V sources, build them with glslc, and include them at build
time.

This fixes all the validation errors I'm aware of, which is nice.
There's almost certainly still things wrong with the Vulkan backend, but
it works well enough for me.

The #include shennanigans is a bit ugly, but my preference (linked
them in directly, using either ld, objcopy, or xxd) is marred by not
having any cross-platform compatibility. (xxd should work, but the
variable name is dependent on the whole path to the file — Yuck! — and
so messes even further with the build system.)