ck_act: Add variable type-checking on debug builds
The CK_VAR system supports several different types of variables (int,
string, action), but uses a single pointer-width integer to store them,
with type information provided by the accessor. This works well — and is
very efficient — but gives terrible crash-y errors if the wrong type is
used.

Add an optional CK_VAR_TYPECHECK option, which will wrap variable values
in a simple structure which includes the type. This is enabled by
default in debug builds, and will quit with a much nicer error message.

This will also give us the option of doing things with variables later
on (such as enumerating over them with some hope of understanding their
contents), which could be useful in utilities or debug builds.