id_str: Tokens always point to raw file data
Currently, the "value" of tokens are stored in an allocated string from
the tempArena. This is ugly, as the tempArena may get quite large, and
we are allocating and copying all sorts of unnecessary things around.

Instead, let's just store the start and length of the token in the
original file data (indeed, we're already doing that, but we'll do it
separately for now, just to ease into the refactoring). We move the
conversion of the value to some type (and hence the parsing of things
like string literals or integers) to separate STR_GetStringValue(), etc.
functions. This means that STR_Tokens can be passed around without ever
needing to allocate anything extra. Only at the point where that's
actually necessary, do we need to allocate for the final value (if we
need to at all), and that ultimately can be done from the allocator
which will use it.

This basically is described in Crafting Interpreters (Robert Nystrom),
Chapter 16.2.

Note that we're not using this brilliantly yet, but we will…