id_str: Get rid of TempArena
Ding, dong, the witch is dead!

Finally, the ugly hack which was the TempArena in STR_ParserState is
gone. This means that the STR_GetString() and STR_GetIdent() functions
now take a pointer to a buffer into which to write the value. This is a
big uglier to use, but allows us to avoid several copies of the string,
which is great for both memory usage and wasted CPU cycles.

Exprect huge stack sizes from some functions, though, particularly with
such a big maximum token size.