ck_var: Allow referencing integer values indirectly
This allows a variable to be set to have the same value as another
variable, and more importantly, the integers _within action definitions_
to use variable values.

Basically, anywhere you'd use an integer, you can instead put @varName
and the value of varName will be used instead.

Note that varName must have been set _before_ it is used.

The goal is to use sprite names rather than numbers in the ACTION files,
e.g.:

	%int SPR_KEENSTANDR 130
	%int SPR_KEENSTANDL 138

	%action CK_ACT_keenStanding 0x098C @SPR_KEENSTANDL @SPR_KEENSTANDR UnscaledFrame 0 1 4 0 32 CK_KeenStandingThink CK_KeenColFunc CK_KeenDrawFunc CK_ACT_keenStanding