vl_sdl2: Use SDL_SetTextureScaleMode() when available.
As pointed out by sjr on the SDL Discourse[1], SDL 2.0.12 introduced the
SDL_SetTextureScaleMode() function, which we can use instead of manually
setting and resetting hints around texture creation. This is much nicer,
but we wrap it around an SDL_VERSION_ATLEAST() so we still have a chance
of compiling against older SDL versions.

[1]: https://discourse.libsdl.org/t/scaling-question/39381/8