id_vl: Handle semitransparent mask tricks
Some Keen mods have sprites which use a non-black (0) colour alongside a
0 mask. This produces a "semitransparent" effect, as the colour is
unconditionally OR-ed with the destination pixel. (If it's zero, and the
mask is also zero, this results in the pixel not being updated. If the
mask is 1, the pixel is first cleared to black, so the colour is being
set.)

Omnispeak previously would never update the pixel value if the mask was
zero, breaking this effect on ID_VL backends which used the PAL8 helpers
(i.e., everything except DOS/EGA).

Implement this properly instead, and restore the effects.