vl_dos: Move both pages at once in SVGA-compat scroll (Fixes #33)
This hopefully is the fix for #33.

Basically, by doing the SVGA-compat blit one page at a time, we
sometimes couldn't maintain the correct "gap" between pages,
particularly if one page was scrolled horizontally, and the other
vertically. After a while, the gap would shrink, and we'd end up in
trouble, with the pages overlapping.

By blitting both pages at once (and keeping them adjacent to each
other), we're guaranteed they'll never overlap.

Also, make the pointers volatile, as that might've been a problem.