id_fs: Fix creating files on Win32
When trying to open a file for writing, we used the "wbx" options to
fopen(). Except 'x' isn't a valid option at all. Wine actually
complained about this (but opened the file succesfully regardless), but
it caused problems on some windows setups. We also used "wbx" in
FS_IsDirWritable(), which resulted in no directory being writable, which
caused the Win32 version (on newer windows versions) defaulting to using
the XDG/SDL_GetPrefPath() fallback instead of the current directory.

NOTE THAT THIS DOES CHANGE THE DEFAULT LOCATION FOR SAVEGAMES AND
CONFIGS ON SOME WINDOWS SETUPS.