id_us_textscreen: Support B8000 text mode in Win32
Add an implementation of the B8000 text mode emulation on top of Win32
console functions.

This works pretty well, except for two major issues:
1. Windows actually doesn't allocate us a console, because we're a
   graphical application.
2. Wine does display it, but has some bugs when displaying from a curses
   terminal. (It mostly works fine when done from a "user" terminal, so
   this is probably a Wine issue.)

The possible fixes are to either build Omnispeak as a Console
Application, instead of a GUI application (which will cause a console
window to appear on boot), or possibly to AllocConsole() ourselves.

We do the former for DEBUG builds, and the latter is behind a config
option, win32_consoleMode

This can be set to any of the following:
- Never — never display the end text screen on Windows.
- IfAvailable — display the end text screen if a console already exists,
  otherwise don't.
- PreferExisting — if a console already exists, display the end text
  screen there, otherwise create a new one.
- CreateNew — always create a new console for the end text screen.

Of course, the console will disappear practically instantly, so it's not
very useful, but should work!