Use FixRefreshBuffers() in a bunch more places
This should handle most cases, though there's possibly a problem where
some parts of the code have the wrong buffer active. So we do need to
work around this properly.

Also reversed the SyncBuffers() implementations to copy from the
previously-rendered frame, which should be pristine. Clearly it isn't
always, though...