id_sd: Add SD_WaitTick()
This uses (on SDL-based backends) a Condition Variable to signal the
game when a new tick has occurred. This basically stops us from busy
waiting, which chewed up a lot of CPU time for no reason. Instead, we
let the OS know our thread is waiting on the timer thread.

Can be disabled with "sd_sdl_waitTicksSpin" in OMNISPK.CFG, which will
revert to the old behaviour when true. (This was mostly because
/AUDIOSYNC is really, really stuttery, and I wondered if this fixed it.
It doesn't, but it's still there to help debug.)

Also add the timer sleep improvement to the opl2alsa version.