id_in: id_us_1: Save/Load Keyboard controls from OMNISPK.CFG (fixes #53)
We now save and load the keyboard bindings from OMNISPK.CFG, which
includes the new options (like QuickSave and QuickLoad), which
previously weren't saved anywhere.

By default, the bindings configured in CONFIG.CKx still take preference:
Omnispeak will save and load what it can from there. This can be
controlled by the "in_preferOmnispeakKeyConfig" option:
- if "false", Omnispeak will still load from OMNISPK.CFG if no
  CONFIG.CKx is found, and will still save the "new" options to
  OMNISPK.CFG. The old options will not be saved.
- if "true", Omnispeak will load from both OMNISPK.CFG and CONFIG.CKx,
  with values from OMNISPK.CFG taking preference. All bindings will be
  saved to OMNISPK.CFG.
- if left unset, the default is "false", except all options are written
  to OMNISPK.CFG (even if the CONFIG.CKx versions will override them on
  load)

Note that the values for the different options (all starting "in_kbd_")
are the raw integer scancodes. Maybe in the future, I'll make it a more
pleasant-to-use enum. But for now, this'll do.

This fixes the issue (#53) reported by Roobar in:
https://pckf.com/viewtopic.php?p=106644#p106644