vl: Calculate a better default window size
For SDL2-based backends, the default window size was always a 3× scaled
version of the game resolute (+ appropriate overscan borders). This was
ideal for my screen.

Instead, let's calculate an ideal scale factor, based on the availabe
space on the desktop.

Ideally, we'd handle multi-monitor cases better, as well as handle the
title bar (and other non-client areas) better. But this is an
improvement.