id_sd_sdl: Support stereo output (by default)
Support outputting to a stereo device, by either duplicating the mono
OPL2 emulator output, or just not mixing-down an OPL3 emulator output.

This hopefully will improve compatibility for some setups which don't
support (or need to do their own channel matrixing for) mono outputs.
(For example, on my system, using 44,100 Hz Stereo at 16-bit can be
copied directly to the output buffer, where as PulseAudio/Pipewire needs
to convert stereo.) I've heard rumours that WASAPI isn't hugely fond of
doing this as well.

And if we ever want to support fancy OPL3 features, this is a
prerequisite, too.

Configure it with "audioChannels" in OMNISPK.CFG. Stereo is now default.
Only 1 and 2 are supported. Maybe we'll add 4-channel OPL if we're
really brave. :-)