id_sd_sdl: Add support for SDL_QueueAudio
This change adds a new option, sd_sdl_queueAudio, which will use the
SDL_QueueAudio() function (when available: it was introduced in SDL
2.0.4). This allows us to have audio-sync like accuracy (particularly
with sound effects, which would sound wrong if the buffer size was too
big before with /NOAUDIOSYNC), but also a higher latency, which can help
slower systems (and particularly windows systems when resampling?).

It also finally pushes the OPL emulation to a different thread (the
timer thread), which should help a bit on WASAPI/Windows?

It's enabled by default on Windows, with SDL 2.0.4+. Otherwise disabled
by default, as the callback version is still cleaner under Linux on my
machine.

But do try it (even on Linux) with the sd_sdl_queueAudio option.

This also fixes us to use OPL_WriteBuffered(), and to clamp the
stereo→mono mixdown with nukedopl3 properly.